dev-requirements.txt
isort==4.3.21
tox
wheel
bumpversion
twine
black
flake8
pylint
cython
sphinx
sphinx-rtd-theme
mypy==0.770
click
requests
python-dotenv